Chapter02/bigtable/requirements.txt
google-cloud-bigtable==0.28.1
google-cloud-core==0.28.0
random==latest
